requirements.txt
openai
python-dotenv
streamlit
tiktoken
requests